Fix overflow in N(N-1); sampler weights a la WenZhe

+++ c++/src/mcmc/learning/learner.cc
@@ -88,7 +88,7 @@ void Learner::LoadNetwork(int world_rank, bool allocate_pi) {
   }
 
   // ratio between link edges and non-link edges
-  link_ratio = network.get_num_linked_edges() / ((N * (N - 1)) / 2.0);
+  link_ratio = network.get_num_linked_edges() / (((double)N * (N - 1)) / 2.0);
 
   ppx_per_heldout_edge_ = std::vector<double>(network.get_held_out_size(), 0.0);
